requirements.txt
opencv-python
numpy
Pillow 
ultralytics
torch
efficientnet_pytorch
PyYAML